pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-kube_sre_gym"
version = "0.1.0"
description = "K8s SRE Self-Improving RL Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime
    "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git@v0.2.1",
    # K8s / GKE
    "kubernetes>=28.0.0",
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "requests>=2.25.0",
    # LLM backends
    "huggingface_hub>=0.20.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "kube_sre_gym.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["kube_sre_gym", "kube_sre_gym.server"]
package-dir = { "kube_sre_gym" = ".", "kube_sre_gym.server" = "server" }
